Use size_t instead of int for a variable that is compared to the result of strlen. Fixes a warning about signed and unsigned comparison. patch by Guillaume LECERF, foxcore gmail com
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@26977 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmenu/menu_filesel.c
@@ -242,7 +242,7 @@ static int open_dir(menu_t* menu,char* args) {
     if(dp->d_name[0] == '.' && strcmp(dp->d_name,"..") != 0)
       continue;
     if (menu_chroot && !strcmp (dp->d_name,"..")) {
-      int len = strlen (menu_chroot);
+      size_t len = strlen (menu_chroot);
       if ((strlen (mpriv->dir) == len || strlen (mpriv->dir) == len + 1)
           && !strncmp (mpriv->dir, menu_chroot, len))
         continue;
